refactor(testing,util): move invoke_coroutine_sync & runs_sync ⇒ util.sync
