+++ docs/api/testing.rst
@@ -15,9 +15,6 @@ General Helpers
 
 .. autoclass:: TestCase
 .. autoclass:: SimpleTestResource
-
-.. autofunction:: invoke_coroutine_sync
-.. autofunction:: runs_sync
 .. autofunction:: rand_string
 .. autofunction:: get_unused_port
 .. autofunction:: redirected
