+++ tests/test_media_handlers.py
@@ -8,6 +8,7 @@ import mujson
 import pytest
 import ujson
 
+import falcon
 from falcon import ASGI_SUPPORTED, media, testing
 
 from _util import create_app  # NOQA
@@ -90,7 +91,7 @@ def test_serialization(asgi, func, body, expected):
     args = (body, b'application/javacript')
 
     if asgi:
-        result = testing.invoke_coroutine_sync(handler.serialize_async, *args)
+        result = falcon.invoke_coroutine_sync(handler.serialize_async, *args)
     else:
         result = handler.serialize(*args)
 
@@ -114,7 +115,7 @@ def test_deserialization(asgi, func, body, expected):
         s = BoundedStream(testing.ASGIRequestEventEmitter(body))
         args.insert(0, s)
 
-        result = testing.invoke_coroutine_sync(handler.deserialize_async, *args)
+        result = falcon.invoke_coroutine_sync(handler.deserialize_async, *args)
     else:
         args.insert(0, io.BytesIO(body))
         result = handler.deserialize(*args)
